refactor: Move array and file function tests to separate files
Extract inline test modules from array_functions.rs and file_functions.rs
into dedicated test files under src/vm/tests/. Maintains all 79 tests
(40 array + 39 file) with identical behavior.
