+++ src/vm/tests/mod.rs
@@ -3,3 +3,5 @@ mod math_variadic;
 mod math_errors;
 mod string_functions;
 mod set_functions;
+mod array_functions;
+mod file_functions;
